add expressions
