+++ src/gsExpressions/meas_expr.h
@@ -53,5 +53,9 @@ public:
     void print(std::ostream &os) const { os << "meas("; _G.print(os); os <<")"; }
 };
 
+/// The measure of a geometry map
+template<class T> EIGEN_STRONG_INLINE
+meas_expr<T> meas(const gsGeometryMap<T> & G) { return meas_expr<T>(G); }
+
 }// namespace expr
 }// namespace gismo
